% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_one.R
\name{add_one}
\alias{add_one}
\title{Adds one}
\usage{
add_one(x)
}
\arguments{
\item{x}{A numeric value to be incremented.}
}
\value{
An integer value, which is x plus one.
}
\description{
This function increments a number by one.
}
